+++ src/core/triangulation_manager.cc
@@ -74,7 +74,7 @@ TriangulationManager<dim>::num_levels() const
 }
 
 template <unsigned int dim>
-const Triangulation<dim> &
+const typename TriangulationManager<dim>::Triangulation &
 TriangulationManager<dim>::get_triangulation() const
 {
   return triangulation;
@@ -96,14 +96,6 @@ TriangulationManager<dim>::get_triangulation(unsigned int relative_level) const
 }
 
 template <unsigned int dim>
-const std::vector<dealii::GridTools::PeriodicFacePair<
-  typename dealii::Triangulation<dim>::cell_iterator>> &
-TriangulationManager<dim>::get_periodic_face_pairs() const
-{
-  return periodicity_vector;
-}
-
-template <unsigned int dim>
 double
 TriangulationManager<dim>::get_volume() const
 {
@@ -115,24 +107,12 @@ void
 TriangulationManager<dim>::generate_mesh(
   const SpatialDiscretization<dim> &discretization_params)
 {
-  if (discretization_params.type == TriangulationType::Rectangular)
-    {
-      discretization_params.rectangular_mesh.generate_mesh(triangulation);
-      discretization_params.rectangular_mesh.collect_periodic_faces(triangulation,
-                                                                    periodicity_vector);
-    }
-  else if (discretization_params.type == TriangulationType::Spherical)
-    {
-      discretization_params.spherical_mesh.generate_mesh(triangulation);
-    }
-  else if (discretization_params.type == TriangulationType::Custom)
-    {
-      AssertThrow(false, FeatureNotImplemented("Custom Triangulation"));
-    }
-  else
-    {
-      AssertThrow(false, UnreachableCode("Invalid TriangulationType"));
-    }
+  Assert(discretization_params.mesh != nullptr,
+         dealii::ExcMessage("Mesh is uninitialized"));
+
+  discretization_params.mesh->generate_mesh(triangulation);
+  discretization_params.mesh->mark_boundaries(triangulation);
+  discretization_params.mesh->mark_periodic(triangulation);
 
   // TODO: Once we move to a pattern of manual initialization after default construction,
   // call this separately using the user_inputs output directory.
