+++ src/user_inputs/CMakeLists.txt
@@ -4,22 +4,21 @@ set(
   spatial_discretization.cc
   temporal_discretization.cc
   user_input_parameters.cc
+	solve_parameters.cc
 )
 
 set(
   _headers
   ${PROJECT_SOURCE_DIR}/include/prismspf/user_inputs/parameter_base.h
-  ${PROJECT_SOURCE_DIR}/include/prismspf/user_inputs/load_initial_condition_parameters.h
   ${PROJECT_SOURCE_DIR}/include/prismspf/user_inputs/spatial_discretization.h
   ${PROJECT_SOURCE_DIR}/include/prismspf/user_inputs/constraint_parameters.h
   ${PROJECT_SOURCE_DIR}/include/prismspf/user_inputs/miscellaneous_parameters.h
   ${PROJECT_SOURCE_DIR}/include/prismspf/user_inputs/temporal_discretization.h
-  ${PROJECT_SOURCE_DIR}/include/prismspf/user_inputs/nonlinear_solve_parameters.h
   ${PROJECT_SOURCE_DIR}/include/prismspf/user_inputs/user_constants.h
   ${PROJECT_SOURCE_DIR}/include/prismspf/user_inputs/io_parameters.h
   ${PROJECT_SOURCE_DIR}/include/prismspf/user_inputs/nucleation_parameters.h
   ${PROJECT_SOURCE_DIR}/include/prismspf/user_inputs/user_input_parameters.h
-  ${PROJECT_SOURCE_DIR}/include/prismspf/user_inputs/linear_solve_parameters.h
+  ${PROJECT_SOURCE_DIR}/include/prismspf/user_inputs/solve_parameters.h
 )
 
 set(_inst_bases user_input_parameters)
